[#2] refactor: 파일 및 변수명 수정
