+++ src/components/table/styles.ts
